requirements.txt
flask
langchain
langchain-community
python-dotenv
duckduckgo-search
langchainhub
requests
